Move Configuration classes to main project
