+++ LegacySignController/src/Configuration/BatteryMonitorConfiguration.cpp
