+++ LegacySignController/src/Configuration/BluetoothConfiguration.cpp
