+++ LegacySignController/src/Configuration/BluetoothConfiguration.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 #include <ArduinoJson.h>
-#include "JsonUtils.h"
+#include "Utils\JsonUtils.h"
 
 class BluetoothConfiguration
 {
