+++ LegacySignController/src/Configuration/Configuration.h
@@ -1,0 +1,15 @@
+#ifndef CONFIGURATION_H
+#define CONFIGURATION_H
+
+#include "BatteryMonitorConfiguration.h"
+#include "BluetoothConfiguration.h"
+#include "DisplayConfigFactory.h"
+#include "DisplayConfiguration.h"
+#include "PowerLedConfiguration.h"
+#include "StyleConfiguration.h"
+#include "StyleConfigFactory.h"
+#include "StyleDefinition.h"
+#include "SystemConfiguration.h"
+#include "Tm1637DisplayConfiguration.h"
+
+#endif
