+++ LegacySignController/src/Configuration/DisplayConfigFactory.cpp
