+++ LegacySignController/src/Configuration/DisplayConfiguration.cpp
