+++ LegacySignController/src/Configuration/PowerLedConfiguration.cpp
