+++ LegacySignController/src/Configuration/StyleConfigFactory.cpp
