+++ LegacySignController/src/Configuration/StyleConfigFactory.h
