+++ LegacySignController/src/Configuration/StyleConfiguration.cpp
