+++ LegacySignController/src/Configuration/StyleDefinition.cpp
