+++ LegacySignController/src/Configuration/StyleDefinition.h
