+++ LegacySignController/src/Configuration/Tm1637DisplayConfiguration.h
@@ -3,7 +3,7 @@
 
 #include <Arduino.h>
 #include <ArduinoJson.h>
-#include "JsonUtils.h"
+#include "Utils\JsonUtils.h"
 
 class Tm1637DisplayConfiguration {
     public:
