+++ LegacySignController/test/test_ButtonProcessor/ButtonProcessorTests.cpp
@@ -9,7 +9,7 @@
 
 // Reference any needed cpp files directly to avoid pulling in
 // references that can't be resolved with the "fake" Arduino environment.
-#include <DebugUtils.h>
+#include "Utils\DebugUtils.cpp"
 #include <ButtonProcessor.h>
 #include "../TestHelpers/MockButton.cpp"
