--- lib/Configuration/Configuration.h
@@ -1,13 +1,0 @@
-#ifndef CONFIGURATION_H
-#define CONFIGURATION_H
-
-#include "DisplayConfiguration.h"
-#include "DisplayConfigFactory.h"
-#include "StyleConfiguration.h"
-#include "SystemConfiguration.h"
-#include "BluetoothConfiguration.h"
-#include "BatteryMonitorConfiguration.h"
-#include "PowerLedConfiguration.h"
-#include "Tm1637DisplayConfiguration.h"
-
-#endif // CONFIGURATION_H
