--- lib/Configuration/library.json
@@ -1,8 +1,0 @@
-{
-  "name": "Configuration",
-  "version": "1.0.0",
-  "dependencies": {
-    "bblanchon/ArduinoJson": "^7.3.0",
-    "DebugUtils": "*"
-  }
-}
