--- lib/PixelMap/library.json
@@ -1,8 +1,0 @@
-{
-  "name": "PixelMap",
-  "version": "1.0.0",
-  "dependencies": {
-    "DisplayPatternLib": "*",
-    "Configuration": "*"
-  }
-}
